--- samples/core/src/test/java/io/oasp/gastronomy/restaurant/common/builders/P.java
@@ -1,6 +1,0 @@
-package io.oasp.gastronomy.restaurant.common.builders;
-
-public interface P<T> {
-
-  public void apply(T target);
-}
