+++ samples/core/src/test/java/io/oasp/gastronomy/restaurant/general/common/api/builders/OrderPositionEtoBuilder.java
@@ -1,4 +1,4 @@
-package io.oasp.gastronomy.restaurant.common.builders;
+package io.oasp.gastronomy.restaurant.general.common.api.builders;
 
 import java.util.LinkedList;
 import java.util.List;
