+++ samples/core/src/test/java/io/oasp/gastronomy/restaurant/general/common/api/builders/P.java
@@ -1,0 +1,6 @@
+package io.oasp.gastronomy.restaurant.general.common.api.builders;
+
+public interface P<T> {
+
+  public void apply(T target);
+}
