+++ samples/core/src/test/java/io/oasp/gastronomy/restaurant/salesmanagement/dataaccess/impl/dao/DrinkDaoTest.java
@@ -15,7 +15,7 @@ import org.springframework.boot.test.SpringApplicationConfiguration;
 import org.springframework.test.context.web.WebAppConfiguration;
 
 import io.oasp.gastronomy.restaurant.SpringBootApp;
-import io.oasp.gastronomy.restaurant.common.builders.DrinkEntityBuilder;
+import io.oasp.gastronomy.restaurant.general.common.api.builders.DrinkEntityBuilder;
 import io.oasp.gastronomy.restaurant.offermanagement.dataaccess.api.DrinkEntity;
 import io.oasp.gastronomy.restaurant.offermanagement.dataaccess.api.dao.DrinkDao;
 import io.oasp.module.jpa.dataaccess.api.RevisionMetadata;
